+++ flang/lib/Optimizer/OpenMP/CMakeLists.txt
@@ -6,7 +6,6 @@ add_flang_library(FlangOpenMPTransforms
   FunctionFiltering.cpp
   GenericLoopConversion.cpp
   GlobalFiltering.cpp
-  HostOpFiltering.cpp
   MapsForPrivatizedSymbols.cpp
   MapInfoFinalization.cpp
   DeleteUnreachableTargets.cpp
