+++ flang/lib/Optimizer/OpenMP/FunctionFiltering.cpp
@@ -49,8 +49,7 @@ checkDeviceImplementationStatus(omp::OffloadModuleInterface offloadModule) {
     if (!redOp.getByrefElementType())
       return WalkResult::advance();
 
-    auto seqTy =
-        mlir::dyn_cast<fir::SequenceType>(*redOp.getByrefElementType());
+    auto seqTy = dyn_cast<fir::SequenceType>(*redOp.getByrefElementType());
 
     bool isByRefReductionSupported =
         !seqTy || !fir::sequenceWithNonConstantShape(seqTy);
@@ -105,7 +104,7 @@ public:
         SymbolTable::UseRange funcUses = *funcOp.getSymbolUses(op);
         for (SymbolTable::SymbolUse use : funcUses) {
           Operation *callOp = use.getUser();
-          if (auto internalFunc = mlir::dyn_cast<func::FuncOp>(callOp)) {
+          if (auto internalFunc = dyn_cast<func::FuncOp>(callOp)) {
             // Do not delete internal procedures holding the symbol of their
             // Fortran host procedure as attribute.
             internalFunc->removeAttr(fir::getHostSymbolAttrName());
@@ -115,6 +114,15 @@ public:
             internalFunc.setVisibility(mlir::SymbolTable::Visibility::Public);
             continue;
           }
+          // Prevent dispatch table entries pointing to deleted functions
+          // from being removed. This prevents the lowering of any
+          // corresponding fir.dispatch ops from triggering errors. These
+          // fir.dt_entry ops will point to an undefined symbol as a result,
+          // which currently doesn't cause an issue, as fir.dispatch-related ops
+          // are later removed by the host op filtering pass.
+          if (isa<fir::DTEntryOp>(callOp))
+            continue;
+
           // If the callOp has users then replace them with Undef values.
           if (!callOp->use_empty()) {
             SmallVector<Value> undefResults;
