+++ flang/lib/Optimizer/Passes/Pipelines.cpp
@@ -351,7 +351,6 @@ void createHLFIRToFIRPassPipeline(mlir::PassManager &pm,
   if (enableOpenMP != EnableOpenMP::None) {
     pm.addPass(flangomp::createLowerWorkshare());
     pm.addPass(flangomp::createLowerWorkdistribute());
-    pm.addPass(flangomp::createHostOpFilteringPass());
   }
   if (enableOpenMP == EnableOpenMP::Simd)
     pm.addPass(flangomp::createSimdOnlyPass());
@@ -465,11 +464,17 @@ void createDefaultFIRCodeGenPassPipeline(mlir::PassManager &pm,
 
   fir::addFIRToLLVMPass(pm, config);
 
-  // Convert applicable OpenMP stack allocations to shared memory allocations
-  // for GPU targets. This pass must run after any alloca-generating passes to
-  // ensure all are adequately accounted for.
-  if (config.EnableOpenMP && !config.EnableOpenMPSimd)
+  if (config.EnableOpenMP && !config.EnableOpenMPSimd) {
+    // Remove all non target-related operations from host functions still
+    // remaining at this point, if compiling for an OpenMP target device. This
+    // is required before translating 'omp' dialect operations to LLVM IR.
+    pm.addPass(mlir::omp::createHostOpFilteringPass());
+
+    // Convert applicable OpenMP stack allocations to shared memory allocations
+    // for GPU targets. This pass must run after any alloca-generating passes to
+    // ensure all are adequately accounted for.
     pm.addPass(mlir::omp::createStackToSharedPass());
+  }
 }
 
 /// Create a pass pipeline for lowering from MLIR to LLVM IR
